thêm logic tính phí vẫn chuyển và lấy địa chỉ tỉnh thành bằng api của ghn

+++ server/src/main/java/org/example/server/entity/Product.java
@@ -12,7 +12,7 @@ public class Product {
     private Long id;
 
     @Column(name = "category_id")
-    private Long categoryId;
+    private Integer  categoryId;
 
     private String name;
 
@@ -30,8 +30,8 @@ public class Product {
     public Long getId() { return id; }
     public void setId(Long id) { this.id = id; }
 
-    public Long getCategoryId() { return categoryId; }
-    public void setCategoryId(Long categoryId) { this.categoryId = categoryId; }
+    public Integer  getCategoryId() { return categoryId; }
+    public void setCategoryId(Integer  categoryId) { this.categoryId = categoryId; }
 
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
