+++ server/src/main/java/org/example/server/repository/ProductRepository.java
@@ -11,7 +11,7 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
     // Lấy 4 sản phẩm đầu tiên để làm "Hàng Hot" trên trang chủ
     List<Product> findTop4ByOrderByIdDesc();
     // Lấy danh sách sản phẩm theo mã Danh mục
-    List<Product> findByCategoryId(Long categoryId);
+    List<Product> findByCategoryId(Integer categoryId);
     // Tìm kiếm sản phẩm theo tên (chứa từ khóa, không phân biệt hoa thường)
     List<Product> findByNameContainingIgnoreCase(String keyword);
 }
